+++ webclient/app/src/app/MainContentRouter.jsx
@@ -1,8 +1,7 @@
-import React from "react";
 import {Route, Routes, useLocation} from "react-router-dom";
 import Navigation from "./commons/navigation/Navigation";
-import ObservationAreaOverview from "./features/observationArea/ObservationAreaOverview";
 import ObservationAreaDetail from "./features/observationArea/ObservationAreaDetail";
+import ObservationAreaOverview from "./features/observationArea/ObservationAreaOverview";
 
 function MainContentRouter() {
     // This needs to be here, because otherwise manual navigation (through `useNavigate()`) does not work properly