+++ webclient/app/src/app/features/observationArea/ObservationAreaDetail.jsx
@@ -3,14 +3,18 @@ import {useNavigate, useParams} from "react-router-dom";
 import {AppBar} from "../../assets/styles/HeaderStyles";
 import ObservationAreaRest from "../../services/ObservationAreaRest";
 import ImageAnnotate from "../imageAnnotate/ImageAnnotate";
+import ObservationVisualization from "../visualizer/ObservationVisualization";
 import ObservationAreaDialog, {MODE as ObservationAreaDialogMode} from "./ObservationAreaDialog";
 import ObservationAreaSelect from "./ObservationAreaSelect";
 
-export default function ObservationAreaDetail() {
+function ObservationAreaDetail(props) {
 
     const {observationAreaId} = useParams();
+    const {showTrajectories = false} = props;
     const navigate = useNavigate();
 
+    const [showTrajectoriesState, setShowTrajectoriesState] = useState(showTrajectories);
+    const [annotateImageSize, setAnnotateImageSize] = useState();
     const [observationAreas, setObservationAreas] = useState();
     const selectedArea = observationAreas?.find(a => String(a.id) === observationAreaId);
     const [editDialogOpen, setEditDialogOpen] = useState(false);
@@ -39,6 +43,10 @@ export default function ObservationAreaDetail() {
         navigate(`/observationarea/${newAreaId}`);
     }
 
+    function onShowTrajectoriesChanged() {
+        setShowTrajectoriesState(!showTrajectoriesState);
+    }
+
     function navigateToHome() {
         navigate("/");
     }
@@ -47,20 +55,35 @@ export default function ObservationAreaDetail() {
         return;
     }
 
-    return (
-        <>
-            <AppBar color="transparent" sx={{boxShadow: "none", left: "0rem", right: "8rem", width: "80%"}}>
+    function renderAppBar() {
+        const appBarSx = {boxShadow: "none", left: "0rem", right: "8rem", width: "80vw", transition: "none"};
+
+        return (
+            <AppBar color="transparent" sx={appBarSx}>
                 <ObservationAreaSelect
                     observationAreas={observationAreas}
                     selectedArea={selectedArea}
                     onHomeClick={navigateToHome}
                     onEditClick={editArea}
                     onAreaChange={navigateToArea}
+                    onShowTrajectoriesChanged={onShowTrajectoriesChanged}
+                    showTrajectories={showTrajectoriesState}
                 />
             </AppBar>
+        )
+    }
+
+
+
+    return (
+        <>
+            {renderAppBar()}
             <ImageAnnotate
                 observationAreaId={observationAreaId}
+                onImageSizeChange={setAnnotateImageSize}
+                sx={{zIndex: 20000}}
             ></ImageAnnotate>
+            {showTrajectoriesState ? <ObservationVisualization streams={selectedArea.saeStreamKeys} imageSize={annotateImageSize}></ObservationVisualization> : null}
             <ObservationAreaDialog
                 open={editDialogOpen}
                 onSubmit={() => setEditDialogOpen(false)}
@@ -70,4 +93,7 @@ export default function ObservationAreaDetail() {
             />
         </>
     )
+
 }
+
+export default ObservationAreaDetail;