+++ webclient/app/src/app/features/observationArea/ObservationAreaSelect.jsx
@@ -1,4 +1,4 @@
-import {Home} from "@mui/icons-material";
+import {Camera, Home} from "@mui/icons-material";
 import EditRoundedIcon from "@mui/icons-material/EditRounded";
 import PlayCircleFilledWhiteIcon from "@mui/icons-material/PlayCircleFilledWhite";
 import StopCircleIcon from "@mui/icons-material/StopCircle";
@@ -6,7 +6,7 @@ import {IconButton, Stack, Tooltip, Typography} from "@mui/material";
 import FormControl from "@mui/material/FormControl";
 import MenuItem from "@mui/material/MenuItem";
 import Select from "@mui/material/Select";
-import React, {useMemo, useState} from "react";
+import {useMemo, useState} from "react";
 import {useTranslation} from "react-i18next";
 import ConfirmationDialog from "../../commons/dialog/ConfirmationDialog";
 import ObservationAreaRest from "../../services/ObservationAreaRest";
@@ -17,7 +17,9 @@ function ObservationAreaSelect(props) {
         selectedArea,
         onHomeClick,
         onEditClick,
-        onAreaChange
+        onAreaChange,
+        onShowTrajectoriesChanged,
+        showTrajectories = false,
     } = props;
 
     const {t} = useTranslation();
@@ -98,7 +100,14 @@ function ObservationAreaSelect(props) {
                     </IconButton>
                 </Tooltip>
             </FormControl>
-
+            <FormControl>
+                <Tooltip title={t('observationArea.showTrajectories')}>
+                    <IconButton sx={{height: "2rem"}} fontSize="small"
+                        onClick={onShowTrajectoriesChanged}>
+                        {showTrajectories ? <Camera color="secondary" /> : <Camera color="primary" />}
+                    </IconButton>
+                </Tooltip>
+            </FormControl>
             <FormControl>
                 <Tooltip title={processingEnabled ? t("observationArea.track.stop.title") : t("observationArea.track.start.title")}>
                     <IconButton sx={{height: "2rem"}}