+++ CHANGELOG.md
@@ -6,6 +6,12 @@ Changelog of observatory-config.
 
 ### Features
 
+-  ab#2266 added trajectories to observation details view ([a39be](https://github.com/starwit/observatory-config/commit/a39be11a3fdb826) Anett Hübner)  
+
+## refs/tags/0.9.8 (2026-06-25)
+
+### Features
+
 -  added card actions to observation area card ([b69e1](https://github.com/starwit/observatory-config/commit/b69e1e2f0574b84) Anett Hübner)  
 
 ## refs/tags/0.9.1 (2025-01-22)